Seeing lots of localhost related test smoke failures
While some of them are 'ports in use' and we can iterate on those later,
most of them seem to be that IPv6 localhost just... doesn't exist,
even though AF_INET6 is ok.  So, try to find localhost from the
appropriate loopback and if it doesn't exist, skip those tests for
sanity.