Document the minimal server listening socket filehandle name
Knowing this is useful in post_setup_hook, or maybe if
one wants to override setup_hook without reimplementing run (ie, the
code in _default_run).

For example, if one wants HTTP::Server::Simple to bind to any free
port, one can override port to specify 0, so that the OS chooses a
free port.  But one then needs to get the port number out somehow and
the obvious method is calling getsockname on the filehandle.

So document this.  It seems unlikely that the minimal implementation
will want or need to change this.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>